requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
astunparse==1.6.3
asyncer==0.0.2
attrs==23.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
billiard==4.2.0
bleach==6.1.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
celery==5.4.0
Django==4.2.9
django-appconf==1.0.6
django-celery-results==2.5.1
django-compressor==4.5.1
django-filter==2.4.0
djangorestframework==3.15.2

